requirements.txt
google-generativeai>=0.7.0
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pypdf>=4.0.0
tiktoken>=0.7.0
pandas>=2.2.0
openpyxl>=3.1.0
tenacity>=8.3.0
python-dotenv>=1.0.0
chardet>=5.2.0
lxml>=5.2.0
